banner-ai-generator/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
starlette==0.27.0

# Authentication & Security
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Dependencies
openai==1.3.8
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
transformers==4.35.2
torch==2.1.1
diffusers==0.24.0
anthropic==0.7.8

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Communication & Message Queue
redis==5.0.1
aioredis==2.0.1
websockets==12.0

# File Processing & Validation
python-magic==0.4.27
# python-magic-bin==0.4.14
Wand==0.6.13
scikit-learn==1.3.2
scipy==1.11.4

# Monitoring & Logging
structlog==23.2.0
colorama==0.4.6

# Async & Concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Development Tools
python-dotenv==1.0.0
watchfiles==0.21.0

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
pytest-cov==4.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Data Processing
pandas==2.1.4
matplotlib==3.8.2

# System Dependencies (Linux/Docker)
# python-magic-bin  # Included above for cross-platform support
